feat: rename file name and ignore config
