+++ lib/browser/client.js
@@ -145,7 +145,7 @@ merge(proto, require('./object'));
 //  */
 // merge(proto, require('./bucket'));
 // multipart upload
-merge(proto, require('./managed_upload'));
+merge(proto, require('./managed-upload'));
 /**
  * Multipart operations
  */
