+++ lib/browser/managed-upload.js
@@ -1,9 +1,11 @@
 
-const fs = require('fs');
+
+// var debug = require('debug')('ali-oss:multipart');
 const is = require('is-type-of');
 const util = require('util');
 const path = require('path');
 const mime = require('mime');
+const copy = require('copy-to');
 
 const proto = exports;
 
@@ -36,8 +38,17 @@ proto.multipartUpload = async function multipartUpload(name, file, options) {
   }
 
   const minPartSize = 100 * 1024;
-  const filename = is.file(file) ? file.name : file;
-  options.mime = options.mime || mime.lookup(path.extname(filename));
+
+  if (!options.mime) {
+    if (is.file(file)) {
+      options.mime = mime.getType(path.extname(file.name));
+    } else if (is.blob(file)) {
+      options.mime = file.type;
+    } else {
+      options.mime = mime.getType(path.extname(file));
+    }
+  }
+
   options.headers = options.headers || {};
   this._convertMetaToHeaders(options.meta, options.headers);
 
@@ -86,6 +97,7 @@ proto.multipartUpload = async function multipartUpload(name, file, options) {
     await options.progress(0, checkpoint, initResult.res);
   }
 
+
   const result = await this._resumeMultipart(checkpoint, options);
   return result;
 };
@@ -104,10 +116,17 @@ proto._resumeMultipart = async function _resumeMultipart(checkpoint, options) {
     file, fileSize, partSize, uploadId, doneParts, name,
   } = checkpoint;
 
+  const internalDoneParts = [];
+
+  if (doneParts.length > 0) {
+    copy(doneParts).to(internalDoneParts);
+  }
+
   const partOffs = this._divideParts(fileSize, partSize);
   const numParts = partOffs.length;
+  let multipartFinish = false;
 
-  const uploadPartJob = function uploadPartJob(self, partNo) {
+  const uploadPartJob = function (self, partNo) {
     return new Promise(async (resolve, reject) => {
       try {
         if (!self.isCancel()) {
@@ -137,9 +156,8 @@ proto._resumeMultipart = async function _resumeMultipart(checkpoint, options) {
   };
 
   const all = Array.from(new Array(numParts), (x, i) => i + 1);
-  const done = doneParts.map(p => p.number);
+  const done = internalDoneParts.map(p => p.number);
   const todo = all.filter(p => done.indexOf(p) < 0);
-
   const defaultParallel = 5;
   const parallel = options.parallel || defaultParallel;
 
@@ -153,7 +171,7 @@ proto._resumeMultipart = async function _resumeMultipart(checkpoint, options) {
     }
   } else {
     // // upload in parallel
-    const jobErr = await this._parallel(todo, parallel, uploadPartJob);
+    const jobErr = await this._parallelNode(todo, parallel, uploadPartJob);
 
     if (this.isCancel()) {
       throw this._makeCancelEvent();
@@ -163,15 +181,23 @@ proto._resumeMultipart = async function _resumeMultipart(checkpoint, options) {
       jobErr[0].message = `Failed to upload some parts with error: ${jobErr[0].toString()} part_num: ${jobErr[0].partNum}`;
       throw jobErr[0];
     }
-  }
 
+    if (this.isCancel()) {
+      jobs = null;
+      throw this._makeCancelEvent();
+    }
+  }
   const result = await this.completeMultipartUpload(name, uploadId, doneParts, options);
   return result;
 };
 
 
-is.file = function (file) {
-  return typeof (File) !== 'undefined' && file instanceof File;
+is.file = function file(obj) {
+  return typeof (File) !== 'undefined' && obj instanceof File;
+};
+
+is.blob = function (blob) {
+  return typeof (Blob) !== 'undefined' && blob instanceof Blob;
 };
 
 /**
@@ -180,7 +206,7 @@ is.file = function (file) {
 proto._getFileSize = async function _getFileSize(file) {
   if (is.buffer(file)) {
     return file.length;
-  } else if (is.file(file)) {
+  } else if (is.blob(file) || is.file(file)) {
     return file.size;
   } if (is.string(file)) {
     const stat = await this._statFile(file);
@@ -239,7 +265,7 @@ WebFileReadStream.prototype._read = function _read(size) {
   size = size || defaultReadSize;
 
   const that = this;
-  this.reader.onload = function (e) {
+  this.reader.onload = function onload(e) {
     that.fileBuffer = new Buffer(new Uint8Array(e.target.result));
     that.file = null;
     that.readFileAndPush(size);
@@ -253,20 +279,22 @@ WebFileReadStream.prototype._read = function _read(size) {
 };
 
 proto._createStream = function _createStream(file, start, end) {
-  if (is.file(file)) {
+  if (is.blob(file) || is.file(file)) {
     return new WebFileReadStream(file.slice(start, end));
-  } else if (is.string(file)) {
-    return fs.createReadStream(file, {
-      start,
-      end: end - 1,
-    });
   }
+  // else if (is.string(file)) {
+  //   return fs.createReadStream(file, {
+  //     start: start,
+  //     end: end - 1
+  //   });
+  // }
+
   throw new Error('_createStream requires File/String.');
 };
 
 proto._getPartSize = function _getPartSize(fileSize, partSize) {
   const maxNumParts = 10 * 1000;
-  const defaultPartSize = 1 * 1024 * 1024;
+  const defaultPartSize = 1024 * 1024;
 
   if (!partSize) {
     return defaultPartSize;
@@ -293,4 +321,12 @@ proto._divideParts = function _divideParts(fileSize, partSize) {
   }
 
   return partOffs;
+};
+
+// cancel is not error , so create an object
+proto._makeCancelEvent = function _makeCancelEvent() {
+  return {
+    status: 0,
+    name: 'cancel',
+  };
 };
