+++ lib/client.js
@@ -132,7 +132,7 @@ merge(proto, require('./object'));
  */
 merge(proto, require('./bucket'));
 // multipart upload
-merge(proto, require('./managed_upload'));
+merge(proto, require('./managed-upload'));
 /**
  * RTMP operations
  */
