+++ lib/common/multipart-copy.js
@@ -169,7 +169,7 @@ proto._resumeMultipartCopy = async function _resumeMultipartCopy(checkpoint, sou
     }
   } else {
     // upload in parallel
-    const errors = await this._parallelCopy(todo, parallel, uploadPartJob, sourceData);
+    const errors = await this._parallelNode(todo, parallel, uploadPartJob, sourceData);
 
     if (this.isCancel()) {
       throw this._makeCancelEvent();
