+++ lib/common/parallel.js
@@ -1,13 +1,7 @@
-'use strict';
 
 const proto = exports;
 
-proto._parallel = async function _parallel(todo, parallel, cb) {
-  const result = await this._parallelCopy(todo, parallel, cb, null);
-  return result;
-};
-
-proto._parallelCopy = async function _parallelCopy(todo, parallel, cb, sourceData) {
+proto._parallelNode = async function _parallelNode(todo, parallel, fn, sourceData) {
   const that = this;
   // upload in parallel
   const jobErr = [];
@@ -22,9 +16,9 @@ proto._parallelCopy = async function _parallelCopy(todo, parallel, cb, sourceDat
     }
 
     if (sourceData) {
-      jobs.push(cb(that, todo[i], sourceData));
+      jobs.push(fn(that, todo[i], sourceData));
     } else {
-      jobs.push(cb(that, todo[i]));
+      jobs.push(fn(that, todo[i]));
     }
 
     if (jobs.length === parallel || (taskIndex === batch && i === (todo.length - 1))) {
