+++ lib/managed-upload.js
@@ -1,11 +1,9 @@
 
-
-// var debug = require('debug')('ali-oss:multipart');
+const fs = require('fs');
 const is = require('is-type-of');
 const util = require('util');
 const path = require('path');
 const mime = require('mime');
-const copy = require('copy-to');
 
 const proto = exports;
 
@@ -38,17 +36,8 @@ proto.multipartUpload = async function multipartUpload(name, file, options) {
   }
 
   const minPartSize = 100 * 1024;
-
-  if (!options.mime) {
-    if (is.file(file)) {
-      options.mime = mime.getType(path.extname(file.name));
-    } else if (is.blob(file)) {
-      options.mime = file.type;
-    } else {
-      options.mime = mime.getType(path.extname(file));
-    }
-  }
-
+  const filename = is.file(file) ? file.name : file;
+  options.mime = options.mime || mime.lookup(path.extname(filename));
   options.headers = options.headers || {};
   this._convertMetaToHeaders(options.meta, options.headers);
 
@@ -97,7 +86,6 @@ proto.multipartUpload = async function multipartUpload(name, file, options) {
     await options.progress(0, checkpoint, initResult.res);
   }
 
-
   const result = await this._resumeMultipart(checkpoint, options);
   return result;
 };
@@ -116,17 +104,10 @@ proto._resumeMultipart = async function _resumeMultipart(checkpoint, options) {
     file, fileSize, partSize, uploadId, doneParts, name,
   } = checkpoint;
 
-  const internalDoneParts = [];
-
-  if (doneParts.length > 0) {
-    copy(doneParts).to(internalDoneParts);
-  }
-
   const partOffs = this._divideParts(fileSize, partSize);
   const numParts = partOffs.length;
-  let multipartFinish = false;
 
-  const uploadPartJob = function (self, partNo) {
+  const uploadPartJob = function uploadPartJob(self, partNo) {
     return new Promise(async (resolve, reject) => {
       try {
         if (!self.isCancel()) {
@@ -156,8 +137,9 @@ proto._resumeMultipart = async function _resumeMultipart(checkpoint, options) {
   };
 
   const all = Array.from(new Array(numParts), (x, i) => i + 1);
-  const done = internalDoneParts.map(p => p.number);
+  const done = doneParts.map(p => p.number);
   const todo = all.filter(p => done.indexOf(p) < 0);
+
   const defaultParallel = 5;
   const parallel = options.parallel || defaultParallel;
 
@@ -171,7 +153,7 @@ proto._resumeMultipart = async function _resumeMultipart(checkpoint, options) {
     }
   } else {
     // // upload in parallel
-    const jobErr = await this._parallel(todo, parallel, uploadPartJob);
+    const jobErr = await this._parallelNode(todo, parallel, uploadPartJob);
 
     if (this.isCancel()) {
       throw this._makeCancelEvent();
@@ -181,23 +163,15 @@ proto._resumeMultipart = async function _resumeMultipart(checkpoint, options) {
       jobErr[0].message = `Failed to upload some parts with error: ${jobErr[0].toString()} part_num: ${jobErr[0].partNum}`;
       throw jobErr[0];
     }
-
-    if (this.isCancel()) {
-      jobs = null;
-      throw this._makeCancelEvent();
-    }
   }
+
   const result = await this.completeMultipartUpload(name, uploadId, doneParts, options);
   return result;
 };
 
 
-is.file = function file(obj) {
-  return typeof (File) !== 'undefined' && obj instanceof File;
-};
-
-is.blob = function (blob) {
-  return typeof (Blob) !== 'undefined' && blob instanceof Blob;
+is.file = function (file) {
+  return typeof (File) !== 'undefined' && file instanceof File;
 };
 
 /**
@@ -206,7 +180,7 @@ is.blob = function (blob) {
 proto._getFileSize = async function _getFileSize(file) {
   if (is.buffer(file)) {
     return file.length;
-  } else if (is.blob(file) || is.file(file)) {
+  } else if (is.file(file)) {
     return file.size;
   } if (is.string(file)) {
     const stat = await this._statFile(file);
@@ -265,7 +239,7 @@ WebFileReadStream.prototype._read = function _read(size) {
   size = size || defaultReadSize;
 
   const that = this;
-  this.reader.onload = function onload(e) {
+  this.reader.onload = function (e) {
     that.fileBuffer = new Buffer(new Uint8Array(e.target.result));
     that.file = null;
     that.readFileAndPush(size);
@@ -279,22 +253,20 @@ WebFileReadStream.prototype._read = function _read(size) {
 };
 
 proto._createStream = function _createStream(file, start, end) {
-  if (is.blob(file) || is.file(file)) {
+  if (is.file(file)) {
     return new WebFileReadStream(file.slice(start, end));
+  } else if (is.string(file)) {
+    return fs.createReadStream(file, {
+      start,
+      end: end - 1,
+    });
   }
-  // else if (is.string(file)) {
-  //   return fs.createReadStream(file, {
-  //     start: start,
-  //     end: end - 1
-  //   });
-  // }
-
   throw new Error('_createStream requires File/String.');
 };
 
 proto._getPartSize = function _getPartSize(fileSize, partSize) {
   const maxNumParts = 10 * 1000;
-  const defaultPartSize = 1024 * 1024;
+  const defaultPartSize = 1 * 1024 * 1024;
 
   if (!partSize) {
     return defaultPartSize;
@@ -321,12 +293,4 @@ proto._divideParts = function _divideParts(fileSize, partSize) {
   }
 
   return partOffs;
-};
-
-// cancel is not error , so create an object
-proto._makeCancelEvent = function _makeCancelEvent() {
-  return {
-    status: 0,
-    name: 'cancel',
-  };
 };
